requirements.txt
alembic==1.4.3
click==7.1.2
Django==2.2
django-autofixture==0.12.1
factory-boy==3.2.0
Faker==5.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
Werkzeug==1.0.1
